+++ precompiles/werc20/werc20.go
@@ -1,11 +1,13 @@
 package werc20
 
 import (
-	"embed"
+	"bytes"
 	"slices"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/core/vm"
+
+	_ "embed"
 
 	ibcutils "github.com/cosmos/evm/ibc"
 	cmn "github.com/cosmos/evm/precompiles/common"
@@ -15,20 +17,17 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-// abiPath defines the path to the WERC-20 precompile ABI JSON file.
-const abiPath = "abi.json"
-
 var (
 	// Embed abi json file to the executable binary. Needed when importing as dependency.
 	//
 	//go:embed abi.json
-	f   embed.FS
+	f   []byte
 	ABI abi.ABI
 )
 
 func init() {
 	var err error
-	ABI, err = cmn.LoadABI(f, abiPath)
+	ABI, err = abi.JSON(bytes.NewReader(f))
 	if err != nil {
 		panic(err)
 	}