+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### IMPROVEMENTS
 
+- [\#758](https://github.com/cosmos/evm/pull/758) Cleanup precompiles abi.json.
+
 ### FEATURES
 
 ### BUG FIXES